Wire T0263 juan 10 merit benefits into selector
